prince-stitcher/pyproject.toml
[tool.poetry]
name = "image-stitcher"
version = "0.1.0"
description = ""
authors = ["Isaac Williams <isaac.andrew.williams@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "^23.2.1"
pyimagej = "^1.4.1"
pydantic = "^2.6.4"
pyyaml = "^6.0.1"
opencv-python-headless = "^4.9.0.80"
s3fs = "^2024.3.1"
watchfiles = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
flake8-bugbear = "^24.2.6"
isort = "^5.13.2"
mypy = "^1.9.0"
pytest = "^8.1.1"


[tool.poetry.group.example.dependencies]
httpx = "^0.27.0"


[tool.isort]
profile = "black"


[tool.mypy]
plugins = "pydantic.mypy"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
